move logo code
